[RFT] Move out constant package from dto package
