+++ src/main/java/net/ow/movie/tmdb/dto/search/SearchResultDTO.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import java.math.BigDecimal;
 import lombok.Data;
-import net.ow.movie.tmdb.dto.constant.MediaType;
+import net.ow.movie.tmdb.constant.MediaType;
 
 @Data
 @JsonTypeInfo(
